refactor: rename `color` dir to `theme`
