+++ apps/readest-app/src/__tests__/components/HighlightColorsEditor.test.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { describe, it, expect, vi, afterEach } from 'vitest';
 import { render, screen, cleanup, fireEvent } from '@testing-library/react';
 
-import HighlightColorsEditor from '@/components/settings/color/HighlightColorsEditor';
+import HighlightColorsEditor from '@/components/settings/theme/HighlightColorsEditor';
 import { HIGHLIGHT_COLOR_HEX } from '@/services/constants';
 import type { DefaultHighlightColor, HighlightColor, UserHighlightColor } from '@/types/book';
