+++ apps/readest-app/src/components/settings/theme/CodeHighlightingSettings.tsx
